Rename TransactionDetails module from vote to stake
