+++ src/modules/transaction/components/TransactionDetails/styles.css
@@ -100,7 +100,7 @@
     text-align: left;
   }
 
-  &.voteValue {
+  &.stakeValue {
     @mixin contentLargest;
 
     display: flex;
@@ -158,13 +158,6 @@
   align-content: flex-start;
   width: 100%;
   flex-wrap: wrap;
-}
-
-.voteTag {
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  align-items: flex-start;
 }
 
 .walletInfo {
